Misscuite-main/requirements.txt
future
emoji
tzlocal
lxml
wget
gtts
faker
beautifulsoup4
requests
python-telegram-bot==12.8
sqlalchemy==1.3.20
ffmpeg-python
psycopg2-binary
feedparser
hachoir
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
googletrans
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pybase64
pySmartDL
validators
nekos.py
aiofiles
pyrate-limiter
cachetools
ujson
pretty_errors
TgCrypto
Pyrogram==1.4.16
yt-dlp
youtube_search_python
youtube_search
asyncio
dateparser
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
fontTools
bing_image_downloader
search_engine_parser
pytz
lyricsgenius
tswift
envparse
better_profanity
motor
cloudscraper
pendulum
pykeyboard
aiohttp
youtube_dl
fuzzysearch
Python_ARQ
httpx[http2]
gpytranslate
google-trans-new

